+++ application/views/catalog.php
@@ -1,0 +1,11 @@
+<h2>Weapons</h2>
+{tableweapons}
+
+<h2>Headgear</h2>
+{tablehead}
+
+<h2>Armour</h2>
+{tablearmor}
+
+<h2>Boots</h2>
+{tableboots}